+++ HISTORY.md
@@ -3,6 +3,7 @@
 ## in development:
 * **New Features:**
     * export pen trails as SVG
+    * access pen trails as SVG_Costume: new "pen trails (SVG)" reporter (experimental, hidden in dev)
 * **Notable Changes:**
 * **Notable Fixes:**
 * **Translation Updates:**
@@ -10,6 +11,7 @@
 ### 2019-12-02
 * new dev version
 * objects: export pen trails as SVG (under construction)
+* threads: new "pen trails (SVG)" reporter (experimental, hidden in dev)
 
 ## v5.3.8:
 * **Notable Change:**